refactor(sql): change datatype of created_at and updated_at to unsignbigInteger
