+++ app/Models/File.php
@@ -1,10 +1,13 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
+use App\Traits\FreshTimestampTrait;
+use App\Traits\PrimaryKeyTrait;
 use Illuminate\Database\Eloquent\Model;
 
 class File extends Model
 {
     //
+    use PrimaryKeyTrait, FreshTimestampTrait;
 }
