+++ app/Models/User.php
@@ -2,6 +2,8 @@
 
 namespace App\Models;
 
+use App\Traits\FreshTimestampTrait;
+use App\Traits\PrimaryKeyTrait;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
@@ -9,7 +11,7 @@ use Illuminate\Notifications\Notifiable;
 class User extends Authenticatable
 {
     use Notifiable;
-
+    use PrimaryKeyTrait, FreshTimestampTrait;
     /**
      * The attributes that are mass assignable.
      *
@@ -36,4 +38,5 @@ class User extends Authenticatable
     protected $casts = [
         'email_verified_at' => 'datetime',
     ];
+
 }
