+++ app/Traits/PrimaryKeyTrait.php
@@ -1,2 +1,12 @@
 <?php
-namespace App;
+namespace App\Traits;
+
+use Illuminate\Support\Str;
+
+trait PrimaryKeyTrait {
+    public static function bootPrimaryKeyTrait() {
+        static::creating(function ($model) {
+            $model->{$model->getKeyName()} = (string) Str::uuid();
+        });
+    }
+}
